requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
apscheduler==3.10.4
anthropic==0.25.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.25.1
